--- Debug/source/SD2_board.d
@@ -1,26 +1,0 @@
-source/SD2_board.o source/SD2_board.d: ../source/SD2_board.c \
- ../source/SD2_board.h ../CMSIS/MKL46Z4.h ../CMSIS/core_cm0plus.h \
- ../CMSIS/cmsis_version.h ../CMSIS/cmsis_compiler.h ../CMSIS/cmsis_gcc.h \
- ../CMSIS/system_MKL46Z4.h ../source/SD2_I2C.h ../source/mma8451.h \
- ../drivers/fsl_port.h ../drivers/fsl_common.h \
- ../CMSIS/fsl_device_registers.h ../CMSIS/MKL46Z4.h \
- ../CMSIS/MKL46Z4_features.h ../drivers/fsl_clock.h ../drivers/fsl_gpio.h \
- ../drivers/fsl_clock.h ../board/pin_mux.h
-../source/SD2_board.h:
-../CMSIS/MKL46Z4.h:
-../CMSIS/core_cm0plus.h:
-../CMSIS/cmsis_version.h:
-../CMSIS/cmsis_compiler.h:
-../CMSIS/cmsis_gcc.h:
-../CMSIS/system_MKL46Z4.h:
-../source/SD2_I2C.h:
-../source/mma8451.h:
-../drivers/fsl_port.h:
-../drivers/fsl_common.h:
-../CMSIS/fsl_device_registers.h:
-../CMSIS/MKL46Z4.h:
-../CMSIS/MKL46Z4_features.h:
-../drivers/fsl_clock.h:
-../drivers/fsl_gpio.h:
-../drivers/fsl_clock.h:
-../board/pin_mux.h:
