--- Debug/source/key.d
@@ -1,14 +1,0 @@
-source/key.o source/key.d: ../source/key.c ../source/key.h \
- ../source/SD2_board.h ../CMSIS/MKL46Z4.h ../CMSIS/core_cm0plus.h \
- ../CMSIS/cmsis_version.h ../CMSIS/cmsis_compiler.h ../CMSIS/cmsis_gcc.h \
- ../CMSIS/system_MKL46Z4.h ../source/SD2_I2C.h ../source/mma8451.h
-../source/key.h:
-../source/SD2_board.h:
-../CMSIS/MKL46Z4.h:
-../CMSIS/core_cm0plus.h:
-../CMSIS/cmsis_version.h:
-../CMSIS/cmsis_compiler.h:
-../CMSIS/cmsis_gcc.h:
-../CMSIS/system_MKL46Z4.h:
-../source/SD2_I2C.h:
-../source/mma8451.h:
