--- Debug/source/mma8451.d
@@ -1,21 +1,0 @@
-source/mma8451.o source/mma8451.d: ../source/mma8451.c \
- ../source/mma8451.h ../drivers/fsl_i2c.h ../drivers/fsl_common.h \
- ../CMSIS/fsl_device_registers.h ../CMSIS/MKL46Z4.h \
- ../CMSIS/core_cm0plus.h ../CMSIS/cmsis_version.h \
- ../CMSIS/cmsis_compiler.h ../CMSIS/cmsis_gcc.h ../CMSIS/system_MKL46Z4.h \
- ../CMSIS/MKL46Z4_features.h ../drivers/fsl_clock.h ../drivers/fsl_port.h \
- ../drivers/fsl_gpio.h
-../source/mma8451.h:
-../drivers/fsl_i2c.h:
-../drivers/fsl_common.h:
-../CMSIS/fsl_device_registers.h:
-../CMSIS/MKL46Z4.h:
-../CMSIS/core_cm0plus.h:
-../CMSIS/cmsis_version.h:
-../CMSIS/cmsis_compiler.h:
-../CMSIS/cmsis_gcc.h:
-../CMSIS/system_MKL46Z4.h:
-../CMSIS/MKL46Z4_features.h:
-../drivers/fsl_clock.h:
-../drivers/fsl_port.h:
-../drivers/fsl_gpio.h:
