--- Debug/source/ringBuffer.d
@@ -1,3 +1,0 @@
-source/ringBuffer.o source/ringBuffer.d: ../source/ringBuffer.c \
- ../source/ringBuffer.h
-../source/ringBuffer.h:
