+++ Debug/source/subdir.mk
@@ -4,40 +4,19 @@
 
 # Add inputs and outputs from these tool invocations to the build variables 
 C_SRCS += \
-../source/MEF.c \
-../source/SD2_I2C.c \
-../source/SD2_board.c \
-../source/TP3_HUBER.c \
-../source/key.c \
-../source/mma8451.c \
+../source/main.c \
 ../source/mtb.c \
-../source/ringBuffer.c \
-../source/semihost_hardfault.c \
-../source/uart_ringBufer.c 
+../source/semihost_hardfault.c 
 
 C_DEPS += \
-./source/MEF.d \
-./source/SD2_I2C.d \
-./source/SD2_board.d \
-./source/TP3_HUBER.d \
-./source/key.d \
-./source/mma8451.d \
+./source/main.d \
 ./source/mtb.d \
-./source/ringBuffer.d \
-./source/semihost_hardfault.d \
-./source/uart_ringBufer.d 
+./source/semihost_hardfault.d 
 
 OBJS += \
-./source/MEF.o \
-./source/SD2_I2C.o \
-./source/SD2_board.o \
-./source/TP3_HUBER.o \
-./source/key.o \
-./source/mma8451.o \
+./source/main.o \
 ./source/mtb.o \
-./source/ringBuffer.o \
-./source/semihost_hardfault.o \
-./source/uart_ringBufer.o 
+./source/semihost_hardfault.o 
 
 
 # Each subdirectory must supply rules for building sources it contributes
@@ -52,7 +31,7 @@ source/%.o: ../source/%.c source/subdir.mk
 clean: clean-source
 
 clean-source:
-	-$(RM) ./source/MEF.d ./source/MEF.o ./source/SD2_I2C.d ./source/SD2_I2C.o ./source/SD2_board.d ./source/SD2_board.o ./source/TP3_HUBER.d ./source/TP3_HUBER.o ./source/key.d ./source/key.o ./source/mma8451.d ./source/mma8451.o ./source/mtb.d ./source/mtb.o ./source/ringBuffer.d ./source/ringBuffer.o ./source/semihost_hardfault.d ./source/semihost_hardfault.o ./source/uart_ringBufer.d ./source/uart_ringBufer.o
+	-$(RM) ./source/main.d ./source/main.o ./source/mtb.d ./source/mtb.o ./source/semihost_hardfault.d ./source/semihost_hardfault.o
 
 .PHONY: clean-source
