--- Debug/source/uart_ringBufer.d
@@ -1,31 +1,0 @@
-source/uart_ringBufer.o source/uart_ringBufer.d: \
- ../source/uart_ringBufer.c ../source/uart_ringBuffer.h \
- ../source/SD2_board.h ../CMSIS/MKL46Z4.h ../CMSIS/core_cm0plus.h \
- ../CMSIS/cmsis_version.h ../CMSIS/cmsis_compiler.h ../CMSIS/cmsis_gcc.h \
- ../CMSIS/system_MKL46Z4.h ../source/SD2_I2C.h ../source/mma8451.h \
- ../drivers/fsl_uart.h ../drivers/fsl_common.h \
- ../CMSIS/fsl_device_registers.h ../CMSIS/MKL46Z4.h \
- ../CMSIS/MKL46Z4_features.h ../drivers/fsl_clock.h \
- ../source/ringBuffer.h ../drivers/fsl_uart_dma.h ../drivers/fsl_uart.h \
- ../drivers/fsl_dma.h ../drivers/fsl_dmamux.h
-../source/uart_ringBuffer.h:
-../source/SD2_board.h:
-../CMSIS/MKL46Z4.h:
-../CMSIS/core_cm0plus.h:
-../CMSIS/cmsis_version.h:
-../CMSIS/cmsis_compiler.h:
-../CMSIS/cmsis_gcc.h:
-../CMSIS/system_MKL46Z4.h:
-../source/SD2_I2C.h:
-../source/mma8451.h:
-../drivers/fsl_uart.h:
-../drivers/fsl_common.h:
-../CMSIS/fsl_device_registers.h:
-../CMSIS/MKL46Z4.h:
-../CMSIS/MKL46Z4_features.h:
-../drivers/fsl_clock.h:
-../source/ringBuffer.h:
-../drivers/fsl_uart_dma.h:
-../drivers/fsl_uart.h:
-../drivers/fsl_dma.h:
-../drivers/fsl_dmamux.h:
